backend/flask_backend/requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dill==0.3.3
future==0.18.2
gast==0.3.3
google-api-core==1.25.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-vision==2.0.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.16
Markdown==3.3.3
matplotlib==3.3.3
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
promise==2.3
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-datasets==1.2.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.26.0
termcolor==1.1.0
tqdm==4.54.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
gunicorn==19.3.0
tensorflow-hub
flask